requirements.txt
redis==5.0.1
boto3==1.34.44
pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.0
aioredis==2.0.1
